engine, libroach: Change type of len in DB{Slice,String} to size_t
This change updates the type of the DBSlice struct in libroach to
use a size_t to denote length, to match rocksdb::Slice. Since DBString
is a sister struct with similar functionality, it also updates that
struct with the same change.

size_t is guaranteed to be unsigned and at least 32 bits.

Fixes #42720.

Release note: None.

+++ pkg/ccl/storageccl/engineccl/rocksdb.go
@@ -105,7 +105,7 @@ func goToCSlice(b []byte) C.DBSlice {
 	}
 	return C.DBSlice{
 		data: (*C.char)(unsafe.Pointer(&b[0])),
-		len:  C.int(len(b)),
+		len:  C.size_t(len(b)),
 	}
 }
 
@@ -121,7 +121,7 @@ func cStringToGoString(s C.DBString) string {
 	if s.data == nil {
 		return ""
 	}
-	result := C.GoStringN(s.data, s.len)
+	result := C.GoStringN(s.data, C.int(s.len))
 	C.free(unsafe.Pointer(s.data))
 	return result
 }
